+++ packages/node-utils/loggerMiddleware.js
@@ -2,48 +2,61 @@ import chalk from 'chalk';
 
 import { logger } from './logger';
 
+function rangeFormat(value, colorMap) {
+  return Object.entries(colorMap).find(([v]) => v === 'default' || value < Number(v))[1];
+}
+
 /**
  * Koa middleware for logging requests using the Appsemble logger.
  */
 export default function loggerMiddleware() {
   return async (ctx, next) => {
+    const { href, res } = ctx;
     const start = Date.now();
     const method = chalk.bold(ctx.method);
-    logger.info(`${method} ${ctx.href} — ${chalk.white(ctx.ip)}`);
+    logger.info(`${method} ${href} — ${chalk.white(ctx.ip)}`);
 
-    function message(status, msg) {
-      const s = `${status} ${msg}`;
-      let duration = Date.now() - start;
-      if (duration < 100) {
-        duration = chalk.green(`${duration}ms`);
-      } else if (duration < 1000) {
-        duration = chalk.yellow(`${duration}ms`);
-      } else {
-        duration = chalk.red(`${duration}ms`);
-      }
-      const format = st => `${method} ${ctx.href} ${st} ${duration}`;
-      if (status < 300) {
-        logger.info(format(chalk.green(s)));
-      } else if (status < 400) {
-        const { location } = ctx.response.headers;
-        if (location) {
-          logger.info(format(chalk.cyan(`${s} → ${location}`)));
-        } else {
-          logger.info(format(chalk.cyan(s)));
-        }
-      } else if (status < 500) {
-        logger.warn(format(chalk.yellow(s)));
-      } else {
-        logger.error(format(chalk.red(s)));
-      }
+    function logResponse() {
+      res.removeListener('finish', logResponse);
+      res.removeListener('close', logResponse);
+
+      const { message, status } = ctx;
+      const s = `${status} ${message}`;
+      const { location } = ctx.response.headers;
+      const duration = Date.now() - start;
+
+      const formatDuration = rangeFormat(duration, {
+        100: chalk.green,
+        1000: chalk.yellow,
+        default: chalk.red,
+      });
+
+      const formatStatus = rangeFormat(status, {
+        200: chalk.red,
+        300: chalk.green,
+        400: chalk.cyan,
+        500: chalk.yellow,
+        default: chalk.red,
+      });
+
+      const level = rangeFormat(status, {
+        200: 'error',
+        400: 'info',
+        500: 'warn',
+        default: 'error',
+      });
+
+      logger.log(
+        level,
+        `${method} ${href} ${formatStatus(location ? `${s} → ${location}` : s)} ${formatDuration(
+          `${duration}ms`,
+        )}`,
+      );
     }
 
-    try {
-      await next();
-    } catch (error) {
-      message(500, 'Internal Server Error');
-      throw error;
-    }
-    message(ctx.status, ctx.message);
+    res.once('finish', logResponse);
+    res.once('close', logResponse);
+
+    return next();
   };
 }