+++ packages/node-utils/loggerMiddleware.test.js
@@ -1,34 +1,42 @@
 import { logger } from '@appsemble/node-utils';
+import { createInstance } from 'axios-test-instance';
 import chalk from 'chalk';
 import Koa from 'koa';
 import lolex from 'lolex';
-import supertest from 'supertest';
 
 import loggerMiddleware from './loggerMiddleware';
+
+class TestError extends Error {}
 
 let app;
 let clock;
 let request;
 
-beforeEach(() => {
+beforeEach(async () => {
   jest.spyOn(logger, 'info').mockImplementation(() => {});
   jest.spyOn(logger, 'log').mockImplementation(() => {});
-  jest.spyOn(logger, 'warn').mockImplementation(() => {});
-  jest.spyOn(logger, 'error').mockImplementation(() => {});
   clock = lolex.install();
   app = new Koa();
-  app.use((ctx, next) => {
+  app.use(async (ctx, next) => {
     Object.defineProperty(ctx.request, 'origin', {
       value: 'https://example.com:1337',
     });
-    return next();
+    try {
+      await next();
+    } catch (error) {
+      if (!(error instanceof TestError)) {
+        throw error;
+      }
+      ctx.status = 400;
+    }
   });
   app.use(loggerMiddleware());
   app.silent = true;
-  request = supertest(app.callback());
+  request = await createInstance(app, { maxRedirects: 0 });
 });
 
-afterEach(() => {
+afterEach(async () => {
+  await request.close();
   clock.uninstall();
 });
 
@@ -45,7 +53,8 @@ it('should log success responses as info', async () => {
     ctx.status = 200;
   });
   await request.get('/fries');
-  expect(logger.info).toHaveBeenCalledWith(
+  expect(logger.log).toHaveBeenCalledWith(
+    'info',
     `${chalk.bold('GET')} https://example.com:1337/fries ${chalk.green('200 OK')} ${chalk.green(
       '1ms',
     )}`,
@@ -58,7 +67,8 @@ it('should log redirect responses as info', async () => {
     ctx.redirect('/');
   });
   await request.get('/fries');
-  expect(logger.info).toHaveBeenCalledWith(
+  expect(logger.log).toHaveBeenCalledWith(
+    'info',
     `${chalk.bold('GET')} https://example.com:1337/fries ${chalk.cyan(
       '302 Found → /',
     )} ${chalk.green('33ms')}`,
@@ -71,7 +81,8 @@ it('should log bad responses as warn', async () => {
     ctx.status = 400;
   });
   await request.get('/burrito');
-  expect(logger.warn).toHaveBeenCalledWith(
+  expect(logger.log).toHaveBeenCalledWith(
+    'warn',
     `${chalk.bold('GET')} https://example.com:1337/burrito ${chalk.yellow(
       '400 Bad Request',
     )} ${chalk.green('3ms')}`,
@@ -84,7 +95,8 @@ it('should log error responses as error', async () => {
     ctx.status = 503;
   });
   await request.get('/wrap');
-  expect(logger.error).toHaveBeenCalledWith(
+  expect(logger.log).toHaveBeenCalledWith(
+    'error',
     `${chalk.bold('GET')} https://example.com:1337/wrap ${chalk.red(
       '503 Service Unavailable',
     )} ${chalk.green('53ms')}`,
@@ -97,7 +109,8 @@ it('should log long request lengths yellow', async () => {
     ctx.status = 200;
   });
   await request.get('/banana');
-  expect(logger.info).toHaveBeenCalledWith(
+  expect(logger.log).toHaveBeenCalledWith(
+    'info',
     `${chalk.bold('GET')} https://example.com:1337/banana ${chalk.green('200 OK')} ${chalk.yellow(
       '400ms',
     )}`,
@@ -110,7 +123,8 @@ it('should log extremely long request lengths red', async () => {
     ctx.status = 200;
   });
   await request.get('/pepperoni');
-  expect(logger.info).toHaveBeenCalledWith(
+  expect(logger.log).toHaveBeenCalledWith(
+    'info',
     `${chalk.bold('GET')} https://example.com:1337/pepperoni ${chalk.green('200 OK')} ${chalk.red(
       '1337ms',
     )}`,
@@ -126,7 +140,8 @@ it('should log errors as internal server errors and rethrow', async () => {
   });
   await request.get('/taco');
   expect(spy).toHaveBeenCalled();
-  expect(logger.error).toHaveBeenCalledWith(
+  expect(logger.log).toHaveBeenCalledWith(
+    'error',
     `${chalk.bold('GET')} https://example.com:1337/taco ${chalk.red(
       '500 Internal Server Error',
     )} ${chalk.green('86ms')}`,
@@ -140,9 +155,24 @@ it('should append the response length if it is defined', async () => {
     ctx.body = '{}';
   });
   await request.get('/fries');
-  expect(logger.info).toHaveBeenCalledWith(
+  expect(logger.log).toHaveBeenCalledWith(
+    'info',
     `${chalk.bold('GET')} https://example.com:1337/fries ${chalk.green('200 OK')} ${chalk.green(
       '1ms',
     )}`,
+  );
+});
+
+it('should log handled errors correctly', async () => {
+  app.use(() => {
+    clock.tick(15);
+    throw new TestError();
+  });
+  await request.get('potatoes');
+  expect(logger.log).toHaveBeenCalledWith(
+    'warn',
+    `${chalk.bold('GET')} https://example.com:1337/potatoes ${chalk.yellow(
+      '400 Bad Request',
+    )} ${chalk.green('15ms')}`,
   );
 });