+++ packages/server/api/paths/apps.js
@@ -54,7 +54,7 @@ export default {
               style: { contentType: 'text/css' },
               sharedStyle: { contentType: 'text/css' },
               icon: {
-                contentType: 'image/png,image/jpg,image/svg+xml,image/tiff,image/webp',
+                contentType: 'image/png,image/jpg,image/tiff,image/webp',
               },
             },
           },
@@ -175,7 +175,7 @@ export default {
               style: { contentType: 'text/css' },
               sharedStyle: { contentType: 'text/css' },
               icon: {
-                contentType: 'image/png,image/jpg,image/svg+xml,image/tiff,image/webp',
+                contentType: 'image/png,image/jpg,image/tiff,image/webp',
               },
             },
           },
@@ -213,7 +213,6 @@ export default {
           content: {
             'image/png': {},
             'image/jpg': {},
-            'image/svg+xml': {},
             'image/tiff': {},
             'image/webp': {},
           },