+++ packages/studio/components/OrganizationsSettings/OrganizationsSettings.jsx
@@ -1,4 +1,4 @@
-import { Form, Icon, Input, Loader, Modal } from '@appsemble/react-components';
+import { CardFooterButton, Form, Icon, Input, Loader, Modal } from '@appsemble/react-components';
 import { normalize, permissions, roles } from '@appsemble/utils';
 import axios from 'axios';
 import classNames from 'classnames';
@@ -594,23 +594,12 @@ export default class OrganizationsSettings extends Component {
             <FormattedMessage {...messages.removeInviteWarning} />
           </div>
           <footer className="card-footer">
-            {/* eslint-disable-next-line jsx-a11y/anchor-is-valid */}
-            <a
-              className="card-footer-item is-link"
-              onClick={this.onCloseInviteDialog}
-              onKeyDown={this.onCloseInviteDialog}
-              role="button"
-              tabIndex="-1"
-            >
+            <CardFooterButton onClick={this.onCloseInviteDialog}>
               <FormattedMessage {...messages.cancel} />
-            </a>
-            <button
-              className={`card-footer-item button is-danger ${styles.cardFooterButton}`}
-              onClick={this.onRemoveInvite}
-              type="button"
-            >
+            </CardFooterButton>
+            <CardFooterButton color="danger" onClick={this.onRemoveInvite}>
               <FormattedMessage {...messages.removeInvite} />
-            </button>
+            </CardFooterButton>
           </footer>
         </Modal>
 
@@ -634,27 +623,19 @@ export default class OrganizationsSettings extends Component {
             )}
           </div>
           <footer className="card-footer">
-            {/* eslint-disable-next-line jsx-a11y/anchor-is-valid */}
-            <a
-              className="card-footer-item is-link"
-              onClick={this.onCloseDeleteDialog}
-              onKeyDown={this.onCloseDeleteDialog}
-              role="button"
-              tabIndex="-1"
-            >
+            <CardFooterButton onClick={this.onCloseDeleteDialog}>
               <FormattedMessage {...messages.cancel} />
-            </a>
-            <button
-              className={`card-footer-item button is-danger ${styles.cardFooterButton}`}
+            </CardFooterButton>
+            <CardFooterButton
+              color="danger"
               onClick={removingMember === user.id ? this.onLeaveOrganization : this.onRemoveMember}
-              type="button"
             >
               {removingMember === user.id ? (
                 <FormattedMessage {...messages.leaveOrganization} />
               ) : (
                 <FormattedMessage {...messages.removeMember} />
               )}
-            </button>
+            </CardFooterButton>
           </footer>
         </Modal>
       </>